requirements.txt
Django~=4.1.0
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.5,<2.10
drf-spectacular>=0.24.2,<0.25
djangorestframework-gis>=1.0,<1.1
django-rosetta>=0.9.8,<0.10
Pillow
django-cors-headers